Cut down on the number of `#[cfg(feature = "gc")]`s outside the implementation of `[VM]ExternRef`
