+++ crates/runtime/src/table.rs
@@ -2,7 +2,6 @@
 //!
 //! `Table` is to WebAssembly tables what `LinearMemory` is to WebAssembly linear memories.
 
-#[cfg(feature = "gc")]
 use crate::externref::VMExternRef;
 use crate::vmcontext::{VMFuncRef, VMTableDefinition};
 use crate::{SendSyncPtr, Store};
@@ -23,7 +22,6 @@ pub enum TableElement {
     FuncRef(*mut VMFuncRef),
 
     /// An `exrernref`.
-    #[cfg(feature = "gc")]
     ExternRef(Option<VMExternRef>),
 
     /// An uninitialized funcref value. This should never be exposed
@@ -36,16 +34,12 @@ pub enum TableElement {
 #[derive(Copy, Clone, PartialEq, Eq, Debug)]
 pub enum TableElementType {
     Func,
-
-    #[cfg(feature = "gc")]
     Extern,
 }
 
 // The usage of `*mut VMFuncRef` is safe w.r.t. thread safety, this
 // just relies on thread-safety of `VMExternRef` itself.
-#[cfg(feature = "gc")]
 unsafe impl Send for TableElement where VMExternRef: Send {}
-#[cfg(feature = "gc")]
 unsafe impl Sync for TableElement where VMExternRef: Sync {}
 
 impl TableElement {
@@ -64,11 +58,9 @@ impl TableElement {
                 let masked = Strict::map_addr(ptr, |a| a & FUNCREF_MASK);
                 Self::FuncRef(masked.cast())
             }
-            #[cfg(feature = "gc")]
             (TableElementType::Extern, None) => Self::ExternRef(None),
-            #[cfg(feature = "gc")]
             (TableElementType::Extern, Some(ptr)) => {
-                Self::ExternRef(Some(VMExternRef::from_raw(ptr.as_ptr())))
+                Self::ExternRef(VMExternRef::from_raw(ptr.as_ptr()))
             }
         }
     }
@@ -83,9 +75,8 @@ impl TableElement {
             // Functions have no ownership, so defer to the prior method.
             TableElementType::Func => TableElement::from_table_value(ty, ptr),
 
-            #[cfg(feature = "gc")]
             TableElementType::Extern => {
-                Self::ExternRef(ptr.map(|p| VMExternRef::clone_from_raw(p.as_ptr())))
+                Self::ExternRef(ptr.and_then(|p| VMExternRef::clone_from_raw(p.as_ptr())))
             }
         }
     }
@@ -106,7 +97,6 @@ impl TableElement {
                 let tagged = Strict::map_addr(e, |e| e | FUNCREF_INIT_BIT);
                 Some(NonNull::new(tagged.cast()).unwrap().into())
             }
-            #[cfg(feature = "gc")]
             Self::ExternRef(e) => e.map(|e| NonNull::new(e.into_raw()).unwrap().into()),
         }
     }
@@ -126,8 +116,6 @@ impl TableElement {
         match self {
             Self::FuncRef(e) => e.cast(),
             Self::UninitFunc => panic!("Uninitialized table element value outside of table slot"),
-
-            #[cfg(feature = "gc")]
             Self::ExternRef(e) => e.map_or(ptr::null_mut(), |e| e.into_raw()),
         }
     }
@@ -148,14 +136,12 @@ impl From<*mut VMFuncRef> for TableElement {
     }
 }
 
-#[cfg(feature = "gc")]
 impl From<Option<VMExternRef>> for TableElement {
     fn from(x: Option<VMExternRef>) -> TableElement {
         TableElement::ExternRef(x)
     }
 }
 
-#[cfg(feature = "gc")]
 impl From<VMExternRef> for TableElement {
     fn from(x: VMExternRef) -> TableElement {
         TableElement::ExternRef(Some(x))
@@ -195,12 +181,7 @@ fn wasm_to_table_type(ty: WasmRefType) -> TableElementType {
         WasmHeapType::Func | WasmHeapType::Concrete(_) | WasmHeapType::NoFunc => {
             TableElementType::Func
         }
-
-        #[cfg(feature = "gc")]
         WasmHeapType::Extern => TableElementType::Extern,
-
-        #[cfg(not(feature = "gc"))]
-        WasmHeapType::Extern => unreachable!(),
     }
 }
 
@@ -512,10 +493,7 @@ impl Table {
     fn type_matches(&self, val: &TableElement) -> bool {
         match (&val, self.element_type()) {
             (TableElement::FuncRef(_), TableElementType::Func) => true,
-
-            #[cfg(feature = "gc")]
             (TableElement::ExternRef(_), TableElementType::Extern) => true,
-
             _ => false,
         }
     }
@@ -561,8 +539,6 @@ impl Table {
                 dst_table.elements_mut()[dst_range]
                     .copy_from_slice(&src_table.elements()[src_range]);
             }
-
-            #[cfg(feature = "gc")]
             TableElementType::Extern => {
                 // We need to clone each `externref`
                 let dst = dst_table.elements_mut();
@@ -583,8 +559,6 @@ impl Table {
                 // `funcref` are `Copy`, so just do a memmove
                 dst.copy_within(src_range, dst_range.start);
             }
-
-            #[cfg(feature = "gc")]
             TableElementType::Extern => {
                 // We need to clone each `externref` while handling overlapping
                 // ranges
