+++ crates/wasmtime/src/runtime/module/registry.rs
@@ -67,7 +67,6 @@ impl ModuleRegistry {
     }
 
     /// Fetches information about a registered module given a program counter value.
-    #[cfg(feature = "gc")]
     pub fn lookup_module_info(&self, pc: usize) -> Option<&dyn wasmtime_runtime::ModuleInfo> {
         let (module, _) = self.module_and_offset(pc)?;
         Some(module.module_info())
